Split AssetBundleLoader into AssetBundleLoaderTest (in editor) and AssetBundleLoader (useful also outside editor mode)
