+++ unity/Holo/Assets/Editor/ModelPreprocessing/AssetBundleHandling/AssetBundleLoaderTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e91da2ee7793dfc48b1bfc359dd984ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
